# CUDA reduction: allow outputs to have different strides