@ngimel 2020-08-13
@jeffdaily we reverted this PR because it was causting consistent ROCm failures in EmbeddingBag tests https://app.circleci.com/pipelines/github/pytorch/pytorch/199835/workflows/6267d613-5d88-4f1f-a221-607208390a30/jobs/6678621, do you know what might be causing those?