@jeffdaily 2020-08-14
@ngimel I've narrowed down the failure in test_nn.py to this line:
https://github.com/pytorch/pytorch/blob/64a7684219bd6974e35efbe93ad20c480f329ea4/test/test_nn.py#L11006

If I comment it out, the fp16 test `TestNNDeviceTypeCUDA.test_embedding_bag_device_cuda_float16` passes.  Still digging.