@ngimel 2020-08-14
It makes sense that failure is happening in `max` mode, the PR changed `max` reduction which is probably called somewhere. Thanks for looking into it!