@jeffdaily 2020-08-24
@ngimel I was unable to determine why this test was failing for ROCm 3.5, but the good news is that the test is passing for ROCm 3.7.  That release only recently became available and we are in the process of upgrading our CI images to ROCm 3.7.  In the meantime, I believe you could try and resubmit this PR if you were to conditionally skip this failing test.  Since it passes for ROCm 3.7, I suggest adding the following to `torch/testing/_internal/common_utils.py` to allow skipping based on ROCm version.

```diff
diff --git a/torch/testing/_internal/common_utils.py b/torch/testing/_internal/common_utils.py
index 1210acdcfd..66ddadad64 100644
--- a/torch/testing/_internal/common_utils.py
+++ b/torch/testing/_internal/common_utils.py
@@ -333,6 +333,7 @@ TEST_WITH_ASAN = os.getenv('PYTORCH_TEST_WITH_ASAN', '0') == '1'
 TEST_WITH_TSAN = os.getenv('PYTORCH_TEST_WITH_TSAN', '0') == '1'
 TEST_WITH_UBSAN = os.getenv('PYTORCH_TEST_WITH_UBSAN', '0') == '1'
 TEST_WITH_ROCM = os.getenv('PYTORCH_TEST_WITH_ROCM', '0') == '1'
+HIP_VERSION = 0.0 if torch.version.hip is None else float(re.search(r"^\d+\.\d+", torch.version.hip)[0])
 # Enables tests that are slow to run (disabled by default)
 TEST_WITH_SLOW = os.getenv('PYTORCH_TEST_WITH_SLOW', '0') == '1'

@@ -387,6 +388,17 @@ def skipIfRocm(fn):
             fn(*args, **kwargs)
     return wrapper

+def skipIfRocmVersionLt(version):
+    def wrap(fn):
+        @wraps(fn)
+        def wrapped_fn(*args, **kwargs):
+            if TEST_WITH_ROCM and HIP_VERSION < version:
+                raise unittest.SkipTest("test doesn't work with ROCm stack < %s" % version)
+            else:
+                fn(*args, **kwargs)
+        return wrapped_fn
+    return wrap
+
```

Then you can decorate this test with `@skipIfRocmVersionLt(3.7)` after having imported the decorator.  `from torch.testing._internal.common_utils import skipIfRocmVersion`.