@jeffdaily 2020-08-24
I suppose the greater concern is not just skipping a test, but that the functionality enabled by this PR is simply broken for ROCm float16 type in ROCm 3.5, in which case, we would expect additional tests to fail that use the max operator until our CI is updated to ROCm 3.7.