@jeffdaily 2020-09-10
@ngimel please attempt to reland this PR now that ROCm CI images have moved to 3.7.